Update README.md
Update build command to have --rm because -rm is deprecated..
Update to include output of using centos6 tag instead of latest, will still be using latest 6 just not 7
Add alternative run command to include --rm for those that don't want left over containers.ctory
Make command to switch to build user include `su - build` instead of `su build` to make it a login shell. This allows it to switch to the actual home directory of build.